server/requirements.txt
django
djangorestframework
requests
gunicorn
django-cors-headers
influxdb_client
pandas